ref: rename services module to controllers
